NO-ISSUE: move simulated agent helpers to test/util/simagent
Keep test/util free of internal/agent imports so agent device package
tests can use util.NewPollConfig without an import cycle.
